fix: implementa blueprints para modularização das rotas e refatora templates para uso correto das URLs
