+++ templates/auth/cadastro.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Página Cadastro</title>
+    <title>Cadastro - Arquivo Nostalgia</title>
     <link rel="stylesheet" href="../../static/css/auth.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
     <link rel="icon" href="../../static/img/logo_arquivo_nostalgia.png" type="image/x-icon">
@@ -57,7 +57,7 @@
         </form>
 
         <div class="link-entrar">
-            Já tem conta? <a href="{{ url_for('login') }}">Entre aqui.</a>
+            <a href="{{ url_for('auth.login') }}">Já tem conta? Faça login</a>
         </div>
 
     </div>
