+++ templates/auth/login.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Página de Login - Arquivo Nostalgia</title>
+    <title>Login - Arquivo Nostalgia</title>
     <link rel="stylesheet" href="../../static/css/auth.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
 </head>
@@ -67,9 +67,10 @@
         </form>
 
         <div class="links-login">
-            <!-- Link dinâmico para a rota de cadastro -->
-            <a href="{{ url_for('register') }}">Não tem conta? Cadastre aqui.</a>
-            <a href="{{ url_for('forgot_password') }}">Esqueceu sua senha?</a>
+            <!-- Link para cadastro -->
+            <a href="{{ url_for('auth.register') }}">Não tem conta? Cadastre-se</a>
+            <!-- Link para esqueci senha -->
+            <a href="{{ url_for('auth.forgot_password') }}">Esqueci minha senha</a>
         </div>
 
     </div>
