+++ templates/conteudo/filmes.html
@@ -18,7 +18,7 @@
     <header class="cabecalho-principal">
       <div class="logo-container">
         <a
-          href="{{ url_for('index') }}"
+          href="{{ url_for('main.index') }}"
           style="
             text-decoration: none;
             display: flex;
@@ -42,37 +42,37 @@
       </div>
 
       <nav class="menu-icones">
-        <a href="{{ url_for('series') }}" class="item-menu">
+        <a href="{{ url_for('series.index') }}" class="item-menu">
           <i class="fa-solid fa-tv"></i>
           <span>Séries</span>
         </a>
-        <a href="{{ url_for('jogos') }}" class="item-menu">
+
+        <a href="{{ url_for('jogos.index') }}" class="item-menu">
           <i class="fa-solid fa-gamepad"></i>
           <span>Jogos</span>
         </a>
-        
+
         {% if current_user.is_authenticated %}
         <!-- MENU PARA QUEM ESTÁ LOGADO -->
-        <a href="{{ url_for('meus_arquivos') }}" class="item-menu">
+        <a href="{{ url_for('main.meus_arquivos') }}" class="item-menu">
           <i class="fa-regular fa-folder-open"></i>
           <span>Arquivos</span>
         </a>
-        <a href="{{ url_for('perfil') }}" class="item-menu">
+        <a href="{{ url_for('main.perfil') }}" class="item-menu">
           <i class="fa-solid fa-circle-user"></i>
           <span>Perfil</span>
         </a>
-        <!-- Botão de Logout  -->
-        <a href="{{ url_for('logout') }}" class="item-menu">
+        <a href="{{ url_for('auth.logout') }}" class="item-menu">
           <i class="fa-solid fa-right-from-bracket"></i>
           <span>Sair</span>
         </a>
         {% else %}
         <!-- MENU PARA QUEM NÃO ESTÁ LOGADO -->
-        <a href="{{ url_for('meus_arquivos') }}" class="item-menu">
+        <a href="{{ url_for('main.meus_arquivos') }}" class="item-menu">
           <i class="fa-regular fa-folder-open"></i>
           <span>Arquivos</span>
         </a>
-        <a href="{{ url_for('login') }}" class="item-menu">
+        <a href="{{ url_for('auth.login') }}" class="item-menu">
           <i class="fa-solid fa-circle-user"></i>
           <span>Login</span>
         </a>
