+++ templates/conteudo/jogos.html
@@ -18,7 +18,7 @@
     <header class="cabecalho-principal">
       <div class="logo-container">
         <a
-          href="{{ url_for('index') }}"
+          href="{{ url_for('main.index') }}"
           style="
             text-decoration: none;
             display: flex;
@@ -43,41 +43,40 @@
       </div>
 
       <nav class="menu-icones">
-        <a href="{{ url_for('filmes') }}" class="item-menu">
-          <i class="fa-solid fa-clapperboard"></i>
+        <a href="{{ url_for('filmes.index') }}" class="item-menu">
+          <i class="fa-solid fa-film"></i>
           <span>Filmes</span>
         </a>
-        <a href="{{ url_for('series') }}" class="item-menu">
+
+        <a href="{{ url_for('series.index') }}" class="item-menu">
           <i class="fa-solid fa-tv"></i>
           <span>Séries</span>
         </a>
-
+        
         {% if current_user.is_authenticated %}
         <!-- MENU PARA QUEM ESTÁ LOGADO -->
-        <a href="{{ url_for('meus_arquivos') }}" class="item-menu">
+        <a href="{{ url_for('main.meus_arquivos') }}" class="item-menu">
           <i class="fa-regular fa-folder-open"></i>
           <span>Arquivos</span>
         </a>
-        <a href="{{ url_for('perfil') }}" class="item-menu">
+        <a href="{{ url_for('main.perfil') }}" class="item-menu">
           <i class="fa-solid fa-circle-user"></i>
           <span>Perfil</span>
         </a>
-        <!-- Botão de Logout  -->
-        <a href="{{ url_for('logout') }}" class="item-menu">
+        <a href="{{ url_for('auth.logout') }}" class="item-menu">
           <i class="fa-solid fa-right-from-bracket"></i>
           <span>Sair</span>
         </a>
         {% else %}
         <!-- MENU PARA QUEM NÃO ESTÁ LOGADO -->
-        <a href="{{ url_for('meus_arquivos') }}" class="item-menu">
+        <a href="{{ url_for('main.meus_arquivos') }}" class="item-menu">
           <i class="fa-regular fa-folder-open"></i>
           <span>Arquivos</span>
         </a>
-        <a href="{{ url_for('login') }}" class="item-menu">
+        <a href="{{ url_for('auth.login') }}" class="item-menu">
           <i class="fa-solid fa-circle-user"></i>
           <span>Login</span>
         </a>
-
         {% endif %}
       </nav>
     </header>
